# clientshop/src/main/resources/openapi/payments.yaml
openapi: 3.0.3
info:
  title: Payment Service API
  description: Реактивный REST API сервиса платежей
  version: 1.0.0

servers:
  - url: http://localhost:8081
    description: Локальный сервер

tags:
  - name: payments
    description: Операции с платежами и балансами

paths:
  /payments/balance/{userId}:
    get:
      tags:
        - payments
      summary: Получить баланс
      description: |
        Возвращает текущий баланс
      operationId: getBalance
      parameters:
        - name: userId
          in: path
          description: Уникальный идентификатор пользователя
          required: true
          schema:
            type: string
            example: "1"
            minLength: 1
            maxLength: 50
            pattern: '^[a-zA-Z0-9_-]+$'
      responses:
        '200':
          description: Баланс успешно получен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
              examples:
                success:
                  summary: Успешный ответ
                  value:
                    userId: "1"
                    balance: 1500000.00
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_user_id:
                  summary: Некорректный ID пользователя
                  value:
                    error: "VALIDATION_ERROR"
                    message: "Invalid user ID"
                    timestamp: 3457654300000
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                internal_error:
                  summary: Внутренняя ошибка
                  value:
                    error: "INTERNAL_ERROR"
                    message: "An unexpected error occurred"
                    timestamp: 3457654300000

  /payments/pay:
    post:
      tags:
        - payments
      summary: Осуществить платеж
      description:
        Проверяет достаточность суммы и делает платеж.
      operationId: processPayment
      requestBody:
        description: Данные для обработки платежа
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
            examples:
              successful_payment:
                summary: Успешный платеж
                value:
                  userId: "1"
                  amount: 500000.00
                  description: "Success payment for"
      responses:
        '200':
          description: Платеж успешно обработан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
              examples:
                success:
                  summary: Успешный платеж
                  value:
                    userId: "1"
                    amount: 500000.00
                    remainingBalance: 1000000.00
                    status: "SUCCESS"
                    message: "Payment processed successfully"
        '400':
          description: Ошибка обработки платежа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
              examples:
                insufficient_funds:
                  summary: Недостаточно средств
                  value:
                    userId: "2"
                    amount: 1000000.00
                    remainingBalance: 500000.00
                    status: "FAILED"
                    message: "Insufficient funds"
                user_not_found:
                  summary: Пользователь не найден
                  value:
                    userId: "unknown_user"
                    amount: 1000000.00
                    remainingBalance: 0.00
                    status: "FAILED"
                    message: "User not found"
                validation_error:
                  summary: Ошибка валидации
                  value:
                    error: "VALIDATION_ERROR"
                    message: "amount: Amount must be positive"
                    timestamp: 3457654300000
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    PaymentRequest:
      type: object
      required:
        - userId
        - amount
      properties:
        userId:
          type: string
          description: Уникальный идентификатор пользователя
          example: "1"
          minLength: 1
          maxLength: 50
          pattern: '^[a-zA-Z0-9_-]+$'
        amount:
          type: number
          format: decimal
          description: Сумма платежа (должна быть положительной)
          example: 1500000.00
          minimum: 0.01
          maximum: 9999999.99
          multipleOf: 0.01
        description:
          type: string
          description: Описание платежа (необязательное поле)
          example: "Order payment for items"
          maxLength: 255
      example:
        userId: "1"
        amount: 1500000.50
        description: "Order payment for items"

    BalanceResponse:
      type: object
      required:
        - userId
        - balance
      properties:
        userId:
          type: string
          description: Уникальный идентификатор пользователя
          example: "1"
        balance:
          type: number
          format: decimal
          description: Текущий баланс пользователя
          example: 1500000.00
          minimum: 0
      example:
        userId: "1"
        balance: 1500000.00

    PaymentResponse:
      type: object
      required:
        - userId
        - amount
        - remainingBalance
        - status
        - message
      properties:
        userId:
          type: string
          description: Уникальный идентификатор пользователя
          example: "1"
        amount:
          type: number
          format: decimal
          description: Сумма обработанного платежа
          example: 1500000.00
        remainingBalance:
          type: number
          format: decimal
          description: Остаток средств на балансе после операции
          example: 1000000.00
          minimum: 0
        status:
          type: string
          description: Статус обработки платежа
          enum:
            - SUCCESS
            - FAILED
          example: "SUCCESS"
        message:
          type: string
          description: Сообщение о результате операции
          example: "Payment processed successfully"
      example:
        userId: "1"
        amount: 1500000.00
        remainingBalance: 1000000.00
        status: "SUCCESS"
        message: "Payment processed successfully"

    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Код ошибки
          example: "VALIDATION_ERROR"
          enum:
            - VALIDATION_ERROR
            - INTERNAL_ERROR
            - USER_NOT_FOUND
            - INSUFFICIENT_FUNDS
        message:
          type: string
          description: Подробное описание ошибки
          example: "amount: Amount must be positive"
        timestamp:
          type: integer
          format: int64
          description: Время возникновения ошибки (Unix timestamp)
          example: 3457654300000
      example:
        error: "VALIDATION_ERROR"
        message: "amount: Amount must be positive"
        timestamp: 3457654300000